1.1.3

+++ main.pyw
@@ -4,8 +4,6 @@
 import os
 import sys
 import json
-import datetime
-import time
 import locale
 import numpy as np
 
@@ -13,9 +11,10 @@ import config
 import utils
 import vkapi
 
-from window import auth 
-from window import tech_info
-from window import mainwindow
+from ui import auth
+from ui import tech_info
+from ui import mainwindow
+
 
 from PyQt5.QtWidgets import QWidget, QDesktopWidget, QApplication, \
     QMessageBox, QFileDialog, QInputDialog, QStyleFactory
@@ -44,11 +43,10 @@ class Auth(QtWidgets.QMainWindow, auth.Ui_MainWindow):
         super().__init__()
         self.setupUi(self)
         self.setWindowIcon(QIcon(config.IconPath))
-        self.setWindowFlags(Qt.WindowCloseButtonHint |
-                            Qt.WindowMinimizeButtonHint)
-        
+        self.setWindowFlags(Qt.WindowCloseButtonHint | Qt.WindowMinimizeButtonHint)
+
         self.statusBar().showMessage("2fa isn't supported")
-        
+
         self.pushButton.clicked.connect(self.autorizations)
         self.pushButton.setShortcut("Return")
 
@@ -57,36 +55,32 @@ class Auth(QtWidgets.QMainWindow, auth.Ui_MainWindow):
         try:
             login = self.lineEdit.text()
             password = self.lineEdit_2.text()
-            
-            if self.action.isChecked():
-                path_oauth = vkapi.OAUTH_PROXY
-                path_api = vkapi.HOST_API_PROXY
-            else:
-                path_oauth = vkapi.OAUTH
-                path_api = vkapi.HOST_API
+
+            path_api = utils.get_host_api(self.action.isChecked())
+            path_oauth = utils.get_host_oauth(self.action.isChecked())
 
             self.statusBar().showMessage('Loading...')
 
-            r = vkapi.autorization(login, password, 
+            r = vkapi.autorization(login, password,
                 vkapi.client_keys[0][0], vkapi.client_keys[0][1],
                 None, None, path_oauth)
 
             # QMessageBox.about(self, "Message", str(r))
-            
+
             json_str = json.dumps(r)
             resp = json.loads(json_str)
 
             if (resp.get('access_token') != None):
-                
+
                 self.statusBar().showMessage('Done!')
                 access_token = resp['access_token']
 
                 self.statusBar().showMessage('Getting refresh_token')
 
-                getRefreshToken = vkapi.refreshToken(access_token, path_api)         
+                getRefreshToken = vkapi.refreshToken(access_token, path_api)
                 refresh_token = getRefreshToken["response"]["token"]
 
-                DATA = {'access_token': access_token, 'token': refresh_token} 
+                DATA = {'access_token': access_token, 'token': refresh_token}
                 utils.save_json("DATA", DATA)
 
                 #Запуск главного окна
@@ -113,8 +107,7 @@ class TechInfo(QWidget, tech_info.Ui_Form):
         super().__init__()
         self.setupUi(self)
         self.setWindowIcon(QIcon(config.IconPath))
-        self.setWindowFlags(Qt.WindowCloseButtonHint |
-                            Qt.WindowMinimizeButtonHint)
+        self.setWindowFlags(Qt.WindowCloseButtonHint | Qt.WindowMinimizeButtonHint)
 
         self.pushButton_3.clicked.connect(self.exit)
         self.pushButton_3.setShortcut("Return")
@@ -148,7 +141,7 @@ class TechInfo(QWidget, tech_info.Ui_Form):
         self.label_4.setText("Hostname: " + hostname)
 
     @pyqtSlot(str)
-    def set_location(self, location):                                                                                                                                              
+    def set_location(self, location):
         self.label_3.setText("Location: " + location)
 
 
@@ -157,7 +150,7 @@ class TechInfo(QWidget, tech_info.Ui_Form):
 
 # Главное окно приложения         
 class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
-    
+
     def __init__(self):
         super().__init__()
         self.setupUi(self)
@@ -171,6 +164,8 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
         self.action_2.triggered.connect(self.Logout)
         self.action_3.triggered.connect(self.Donate)
         self.action_4.triggered.connect(self.TechInformation)
+        #self.progressBar.setFormat("%{.2f}%".format('p'))
+        self.progressBar.setFormat("%p%")
         self.action_4.setShortcut("Ctrl+T")
 
 
@@ -183,12 +178,8 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
             refresh_token = data_token["token"]
 
             try:
-                if(self.action_5.isChecked()):
-                    path_api = vkapi.HOST_API_PROXY
-                    path_oauth = vkapi.OAUTH_PROXY
-                else:
-                    path_api = vkapi.HOST_API
-                    path_oauth = vkapi.OAUTH
+                path_api = utils.get_host_api(self.action_5.isChecked())
+                path_oauth = utils.get_host_oauth(self.action_5.isChecked())
 
                 data = vkapi.get_audio(refresh_token, path_api)
                 utils.save_json('response.json', data)
@@ -219,7 +210,7 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
 
                     i += 1
 
-                self.label.setText("Всего аудиозаписей: " + str(count_track) 
+                self.label.setText("Всего аудиозаписей: " + str(count_track)
                     + " Выбрано: " + str(0) + " Загружено: " + str(0))
 
             except Exception as e:
@@ -230,41 +221,28 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
 
 
     def Downloads(self):
-        try:
-            if (self.action_7.isChecked()):
-                PATH = utils.get_path(self, QFileDialog)
-
-                if PATH == "":
-                    QMessageBox.warning(self, "Внимание",
-                        "Вы не выбрали путь, поэтому аудиозаписи," + 
-                        "будут загружены в директорию с программой")
-                    PATH = os.getcwd()
-
-                self.label_2.setText("Путь для скачивания: " + PATH)
-
-            else:
-                PATH = os.getcwd()
-                self.label_2.setText("Путь для скачивания: " + PATH)
+        try:  
+            PATH = utils.get_path(self, self.action_7.isChecked(), QFileDialog)
+            self.label_2.setText("Путь для скачивания: " + PATH)
 
             self.completed = 0
             downloads_list = []
             getSelected = self.treeWidget.selectedItems()
-            
+
             for i in getSelected:
                 downloads_list.append(int(i.text(0)))
 
             with open('response.json', encoding='utf-8') as data_json:
                 data = json.loads(data_json.read())
 
-            #self.progressBar.setRange(0, np.size(downloads_list))
             count_track = data['response']['count']
 
-            QApplication.processEvents()
+            #QApplication.processEvents()
 
             if (np.size(downloads_list) == 0):
                     QMessageBox.information(self, "Информация",
                      "Ничего не выбрано.")
-            
+
             self.th = Downloads_file(downloads_list, PATH)
             self.th.progress_range.connect(self.progress)
             self.th.progress.connect(self.progressBar.setValue)
@@ -275,12 +253,12 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
             self.th.start()
 
             #self.label_3.setText("Загружается:3")
-                         
+
         except Exception as e:
             QMessageBox.critical(self, "F*CK", str(e))
 
 
-    @pyqtSlot(int)
+    @pyqtSlot()
     def finished_loader(self):
         QMessageBox.information(self, "Информация", "Аудиозаписи загружены")
 
@@ -288,11 +266,10 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
     def loading_audio(self, song_name):
         self.label_3.setText("Загружается: " + song_name)
 
-
     @pyqtSlot(str)
     def unavailable_audio(self, song_name):
         QMessageBox.warning(self, "Внимание",
-         "Аудиозапись: " + song_name + " недоступна в вашем регионе") 
+         "Аудиозапись: " + song_name + " недоступна в вашем регионе")
 
     @pyqtSlot(int)
     def progress(self, range):
@@ -302,10 +279,10 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
     def AboutMessage(self):
         QMessageBox.about(
             self, "О программе",
-            "<b>" + config.ApplicationName 
+            "<b>" + config.ApplicationName
             + "</b> - " + config.Description + "<br><br><b>Версия: </b>"
             + config.ApplicationVersion
-            + "<br><b>Стадия: </b> " 
+            + "<br><b>Стадия: </b> "
             + config.ApplicationBranch
             + "<br><br>Данное ПО распространяется по лицензии "
             + "<a href=" + config.License +">MIT</a>, исходный код доступен"
@@ -316,7 +293,7 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
     def Donate(self):
         QMessageBox.about(
             self, "Помощь проекту",
-            "<b>Дать разработчику на чай</b>" 
+            "<b>Дать разработчику на чай</b>"
             + "<br><br><b>QIWI: </b> <a href="
             + config.Qiwi
             + ">" + config.Qiwi + "</a>"
@@ -327,22 +304,25 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
 
 
     def TechInformation(self):
-        if self.action_5.isChecked():
-            self.tech_info_window = TechInfo(vkapi.HOST_API_PROXY, vkapi.OAUTH_PROXY)
-        else:
-            self.tech_info_window = TechInfo(vkapi.HOST_API, vkapi.OAUTH)
+        path_api = utils.get_host_api(self.action_5.isChecked())
+        path_oauth = utils.get_host_oauth(self.action_5.isChecked())
+        
+        self.tech_info_window = TechInfo(path_api, path_oauth)
 
 
     def Logout(self):
         reply = QMessageBox.question(self, "Выход из аккаунта",
-         "Вы точно хотите выйти из аккаунта?", 
+         "Вы точно хотите выйти из аккаунта?",
             QMessageBox.Yes | QMessageBox.No, QMessageBox.No)
 
         if reply == QMessageBox.Yes:
-            os.remove("DATA")
+            if (utils.check_file_path("DATA")):
+                os.remove("DATA")
+            else:
+                print("WTF?")
 
             if (utils.check_file_path("response.json")):
-                os.remove("response.json")    
+                os.remove("response.json")
 
             self.auth_window = Auth()
             self.auth_window.show()
@@ -353,7 +333,7 @@ class MainWindow(QtWidgets.QMainWindow, mainwindow.Ui_MainWindow):
 
 class Downloads_file(QThread):
 
-    finished = pyqtSignal(int)
+    finished = pyqtSignal()
     progress_range = pyqtSignal(int)
     progress = pyqtSignal(int)
     loading_audio = pyqtSignal(str)
@@ -375,17 +355,16 @@ class Downloads_file(QThread):
         try:
             with open('response.json', encoding='utf-8') as data_json:
                 data = json.loads(data_json.read())
-            
+
             self.completed = 0
 
             count_track = data['response']['count']
             selected = np.size(self.downloads_list)
 
-            self.progress_range.emit(selected)
 
             for item in self.downloads_list:
                 self.completed += 1
-
+                
                 artist = data['response']['items'][item-1]['artist']
                 title = data['response']['items'][item-1]['title']
 
@@ -396,23 +375,29 @@ class Downloads_file(QThread):
 
                 filename = self.PATH + "/" + utils.remove_symbols(song_name) + ".mp3"
                 url = data['response']['items'][item-1]['url']
-                
+
+                #total = int(utils.get_size_content(url))
+                #self.progress_range.emit(total)
+
                 if (data['response']['items'][item-1]['url'] == ""):
                     self.unavailable_audio.emit(song_name)
-                    
+
                 else:
                     self.message.emit(msg)
                     self.loading_audio.emit(song_name)
-                    utils.downloads_files_in_wget(url, filename)
+                    utils.downloads_files_in_wget(url, filename, self.update_progress)
+                    #utils.downloads_files(url, filename, self.progress)
 
-                self.progress.emit(self.completed)
-                    
-
-            self.finished.emit(1)
+            self.finished.emit()
             self.loading_audio.emit('')
-                
+
         except Exception as e:
             pass
+
+
+    def update_progress(self, current, total, width=80):
+        self.progress_range.emit(total)
+        self.progress.emit(current)
 
 
 class NetworkInfo(QThread):
@@ -422,7 +407,7 @@ class NetworkInfo(QThread):
     hostname = pyqtSignal(str)
     location = pyqtSignal(str)
     api = pyqtSignal(str)
-    oauth = pyqtSignal(str) 
+    oauth = pyqtSignal(str)
 
     def __init__(self, host_api, host_oauth):
         super().__init__()
@@ -433,7 +418,7 @@ class NetworkInfo(QThread):
     def run(self):
         try:
             self.internal_ip.emit(utils.get_internal_ip())
-            
+
             data = utils.get_network_info()
             loc = data['country'] + ", " + data['region'] + ", " + data['city']
 
@@ -441,7 +426,7 @@ class NetworkInfo(QThread):
             self.hostname.emit(data['hostname'])
             self.location.emit(loc)
 
-        except:    
+        except:
             self.internal_ip.emit(utils.get_internal_ip())
             self.external_ip.emit(None)
             self.hostname.emit(None)
@@ -463,6 +448,7 @@ class NetworkInfo(QThread):
 
 
     def __del__(self):
+        print("Never Say Goodbye")
         self.wait()
